refactor: URL Extension, Foldering, JSON MODEL 수정
