+++ WeatherForecast/WeatherForecast/Protocol/API.swift
@@ -1,0 +1,12 @@
+//
+//  API.swift
+//  WeatherForecast
+//
+//  Created by 박태현 on 2021/10/04.
+//
+
+import Foundation
+
+protocol API {
+    var url: String { get }
+}
