+++ WeatherForecast/WeatherForecast/Protocol/Query.swift
@@ -1,0 +1,12 @@
+//
+//  Query.swift
+//  WeatherForecast
+//
+//  Created by 박태현 on 2021/10/04.
+//
+
+import Foundation
+
+protocol Query {
+    var description: String { get }
+}
